remove Mongo class.
this has been deprecated for some time, in favor of creating a Connection
instance and getting a Database instance from the connection. i'm removing
this before the release because i'd rather not have people get confused as
to why it is lying around.

the only real useful functionality left here that isn't recreated elsewhere
is auto-ref and auto-deref. this will probably get added back to the driver
as a SONManipulator at some point.