Python BSON decoder returns a plain dictionary instead of SON.
This matches the behavior of the C decoder (returning SON using the C
decoder is dramatically slower than returning a plain dict). This also hides
from the user the fact that BSON is ordered - internally we use SON when
needed (to make sure that _id is first, to make sure that command verbs are
first), but for user documents it shouldn't matter anyway.