DEPRECATING Cursor.__len__ in favor of Cursor.size
The reason for this deprecation is a bit complex:
list(...) calls _PyObject_LengthHint to guess how much space will be
required for the returned list. That method in turn calls __len__.
Therefore, calling list(...) on a Cursor instance would require at least
two round trips to the database if we keep __len__ - this makes it about
twice as slow as [x for x in Cursor], which isn't obvious to users.
Not defining __len__ here makes performance more consistent