slaveok and safe enhancement PYTHON-181 PYTHON-246
This change adds a new module common.py with some
common functions and classes used in a number of
pymongo modules.

The most important of these is the BaseObject class.
This provides a new base class that Connection, Database,
and Collection inherit from. The class provides attributes
common to each of these subclasses including slave_okay,
safe and getlasterror options.

With this change you can now set slaveok, safe, w, wtimeout,
fsync and j at the Connection, Database, Collection or per-
operation level. Each level inherits from and can override
the previous.

All of these options are now supported as keyword parameters
to connection.Connection and in the mongodb URI.