Add support for BSON Binary and Code objects to json_util
Because Binary and Code extend the built-in str type, the default
callback in JSONEncoder cannot be used to change their representation.
The solution is to instead provide a JSONEncoder subclass that
implements the correct encoding, which can be passed to json.dump and
json.dumps as the cls keyword argument. A JSONDecoder subclass is
provided for consistency.

Signed-off-by: Eric Naeseth <eric@thumbtack.com>