PYTHON-722 Use SocketInfo in a with-statement.
SocketInfo is now responsible for sending and receiving messages,
for closing itself on error, and for returning itself to the pool
when its with-statement is exited.

Pool.discard_socket is removed. Returning a closed socket to the
pool has the same effect as discard_socket had.

Remove client._exhaust_next.