Correct cluster type if a primary uses internal IP.
Begin with CLUSTER_TYPE.ReplicaSetNoPrimary and a
seed list of "A". The client connects to A and finds that
it's a primary, and its host list is ["B"], where "B" is the
internal IP for "A". The client should remove "A" from its
list of servers and add "B" as SERVER_TYPE.Unknown,
and start monitoring "B".

The cluster type should remain ReplicaSetNoPrimary,
since server "B" is of type Unknown, and "A" is no longer
in the ClusterDescription. However, the cluster type
winds up as ReplicaSetWithPrimary.