Run tests even if prior run didn't clean up users.
If a previous test run created a user and didn't remove it (e.g. the test hung or a debugging session was terminated), the next test run will fail because of the narrowed localhost exception (SERVER-12621). Work around this by trying to log in as the test user, if that fails re-raise the original error.