PYTHON-525 Socket timeout for monitoring is connect_timeout.
The spec's justification is that while a client waits for a server to respond to a
connection, the client does not know if the server will respond eventually or if it is
down. Users help the client guess by supplying a reasonable connectTimeoutMS for
their network.

The socketTimeoutMS, on the other hand, must account for both network latency
and the operation's duration. Applications should set a long or infinite
socketTimeoutMS to wait for long-running operations.

A socket used for monitoring does two things: it connects and calls ismaster. Both
operations are fast on the server, so only network latency matters. Thus both
operations use connectTimeoutMS, since that is the value users supply to help the
client guess if a server is down, based on users' knowledge of expected latencies
on their networks.