Fix test_init_disconnected.
It failed when the replica set config contains hostnames that differ from the
DB_IP environment variable that PyMongo's tests use. E.g. if DB_IP is
"127.0.0.1" but the replica set is configured to use hostnames like
"localhost:27017", then test_init_disconnected failed because the tests' global
"host" variable is "127.0.0.1" but MongoClient.host is "localhost".