API: Add `Cursor.pop_bson_left`
This method can be used to retrieve data from MongoDB avoiding
deserialization into a nested list of Python dictionaries.  In my
application this serialisation overhead was expensive, taking ~100ms for
6000 records.  My intention was to forward the data to web client, and
BSON is just as suitable for this as JSON.

This commit exposes the previously private ``BSONList`` class as API.

This is a partial solution to PYTHON-472: "Provide an API for inserting and
returning raw BSON".