Improved int/long encoding and added unsigned int64 encoding support
There was an issue with encoding an unsigned 64-bit int/long. This would occur
when encoding a value larger than (2^63 - 1), which is still a valid int64.
Now, the size of Python ints/longs are intelligently detected and the encoded
BSON is correct. However, storing a unsigned int64 will still be decoded as a
signed int64. It is up to the enduser to properly know whether the value was
originally signed or unsigned and to perform the conversion manually, which
is the standard across the other MongoDB drivers.