Preserve order of fields on JSON
I want to preserve field order on my application.

To do that, mongo has to use with ordered dicts instead of the standard `dict`.
Python's OrderedDict and Mongo's SON are good candidates for that.

As far as I can tell, there are 2 places where I need to make the change:
- On MongoClient's `document_class` (When reading from MongoDB)
- On json_util's `loads` (When parsing a JSON file)

I do both on my application, but it seems like it would be a reasonable default behavior