PYTHON-1358 - Switch to Jira Release Notes links
Jira server no longer allows anonymous users to view
the "roadmap" page for a project release. This broke most of the
"Issues Resolved" links in the changelog. The Jira Release Notes
pages are still viewable by anonymous users.

This change also adds the release notes for 2.9.5 and 3.3.1, which
never made it into master.