Provide a Docstring for MongoClient
MongoClient inherits directly from common.BaseObject without specifying a docstring. This causes the class' docstring to be set as equal as its parent's.

BaseObject is an internal class with a not very friendly docstring that includes the sentence "SHOULD NOT BE USED BY DEVELOPERS EXTERNAL TO MONGODB." which is very misleading to be displayed as MongoClient's docstring.